requirements.txt
aiohttp==3.9.5
aioresponses==0.7.6
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.1
async-timeout==4.0.3
attrs==23.2.0
awscli==1.29.54
backoff==2.2.1
blinker==1.7.0
boto3==1.28.23
botocore==1.31.54
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
Cython==3.0.0
dataclasses-json==0.6.4
defusedxml==0.7.1
distro==1.9.0
docker==6.1.3
docutils==0.16
exceptiongroup==1.2.1
fastapi==0.110.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imageio==2.34.1
inference==0.9.22
iniconfig==2.0.0
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langsmith==0.1.51
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.25.2
onnxruntime==1.15.1
openai==1.23.6
opencv-python==4.8.0.76
opencv-python-headless==4.9.0.80
orjson==3.10.1
packaging==23.2
pandas==2.2.2
pendulum==3.0.0
piexif==1.1.3
pillow==10.3.0
pluggy==1.5.0
prettytable==3.10.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
PuLP==2.8.0
py-cpuinfo==9.0.0
pyarrow==16.0.0
pyasn1==0.6.0
pybase64==1.3.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0b1
Pygments==2.17.2
pyparsing==3.1.2
pytest==8.1.2
pytest-asyncio==0.21.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.6.2
scikit-image==0.23.2
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.1
six==1.16.0
skypilot==0.4.1
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
streamlit==1.33.0
structlog==24.1.0
supervision==0.20.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
thop==0.1.1.post2209072238
tifffile==2024.4.24
time-machine==2.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.3
urllib3==1.26.18
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.9.4
zxing-cpp==2.2.0
